renew entities

+++ news-service/src/main/java/com/ntsoftware/myvspc/newsservice/entities/dao/NewsEntity.java
@@ -36,6 +36,6 @@ public class NewsEntity {
     private UUID imagePreviewId;
 
     @Column(name = "type")
-    private int type;
+    private long type;
 
 }
